Correção do erro para o método de cadastrar

+++ api/api/src/main/java/br/com/ecommerce/api/service/ProdutoService.java
@@ -65,9 +65,13 @@ public class ProdutoService {
     }
 
     // Método para deletar produto
-    public void deletarProduto(Integer id){
+    public Produto deletarProduto(Integer id){
         Produto produto = buscarPorId(id);
+        if(produto == null){
+            return null;
+        }
         produtoRepository.delete(produto);
+        return produto;
     }
 }
